requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==4.1
djangocodemirror==2.1.0
factory-boy==3.2.1
Faker==14.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
redis==4.3.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
wrapt==1.14.1
zope.interface==5.4.0
